[feat]: [CDS-79816]: Add Deployment Freeze Support for Custom Stage Environment Step (#54363)
* [feat]: [CDS-79376]: add freeze helper class

* [feat]: [CDS-79376]: fix checkstyle issues

* [feat]: [CDS-79376]: add freeze support in custom stage plan creator

* [feat]: [CDS-79376]: add freeze support in custom stage plan creator - 2

* [feat]: [CDS-79376]: add freeze support in custom stage plan creator - 3

* [fix]: [CDS-79080]: adding freeze code to service step helper

* [fix]: [CDS-79080]: adding freeze code to service step helper - 2

* [fix]: [CDS-79080]: remove plan creator changes
