+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/environment/steps/CustomStageEnvironmentStep.java
@@ -30,6 +30,7 @@ import io.harness.cdng.service.steps.helpers.beans.ServiceStepV3Parameters;
 import io.harness.cdng.stepsdependency.constants.OutcomeExpressionConstants;
 import io.harness.exception.InvalidRequestException;
 import io.harness.exception.WingsException;
+import io.harness.freeze.beans.FreezeEntityType;
 import io.harness.logging.CommandExecutionStatus;
 import io.harness.logging.LogLevel;
 import io.harness.logging.UnitProgress;
@@ -63,11 +64,13 @@ import io.harness.utils.IdentifierRefHelper;
 import software.wings.beans.LogColor;
 import software.wings.beans.LogHelper;
 
+import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.EnumMap;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -146,6 +149,22 @@ public class CustomStageEnvironmentStep implements ChildrenExecutable<CustomStag
       saveNonVariableOverridesToSweepingOutput(
           ambiance, scopedEnvironmentRef, isOverridesV2enabled, mergedOverrideV2Configs, ngEnvironmentConfig);
 
+      Map<FreezeEntityType, List<String>> entityMap = new HashMap<>();
+      entityMap.put(FreezeEntityType.ORG, Lists.newArrayList(AmbianceUtils.getOrgIdentifier(ambiance)));
+      entityMap.put(FreezeEntityType.PROJECT, Lists.newArrayList(AmbianceUtils.getProjectIdentifier(ambiance)));
+      entityMap.put(FreezeEntityType.PIPELINE, Lists.newArrayList(AmbianceUtils.getPipelineIdentifier(ambiance)));
+      entityMap.put(FreezeEntityType.ENVIRONMENT,
+          Lists.newArrayList(IdentifierRefHelper.getRefFromIdentifierOrRef(environment.get().getAccountId(),
+              environment.get().getOrgIdentifier(), environment.get().getProjectIdentifier(),
+              environment.get().getIdentifier())));
+      entityMap.put(FreezeEntityType.ENV_TYPE, Lists.newArrayList(environment.get().getType().name()));
+
+      ChildrenExecutableResponse childrenExecutableResponse = serviceStepV3Helper.executeFreezePart(
+          ambiance, entityMap, List.of(ENVIRONMENT_STEP_COMMAND_UNIT, OVERRIDES_COMMAND_UNIT));
+      if (childrenExecutableResponse != null) {
+        return childrenExecutableResponse;
+      }
+
       return ChildrenExecutableResponse.newBuilder()
           .addAllLogKeys(emptyIfNull(StepUtils.generateLogKeys(StepUtils.generateLogAbstractions(ambiance),
               List.of(ENVIRONMENT_STEP_COMMAND_UNIT, OVERRIDES_COMMAND_UNIT))))
@@ -196,7 +215,16 @@ public class CustomStageEnvironmentStep implements ChildrenExecutable<CustomStag
     long environmentStepStartTs = AmbianceUtils.getCurrentLevelStartTs(ambiance);
     final List<StepResponse.StepOutcome> stepOutcomes = new ArrayList<>();
 
-    StepResponse stepResponse = SdkCoreStepUtils.createStepResponseFromChildResponse(responseDataMap);
+    final EnvironmentOutcome environmentOutcome = (EnvironmentOutcome) sweepingOutputService.resolve(
+        ambiance, RefObjectUtils.getOutcomeRefObject(OutputExpressionConstants.ENVIRONMENT));
+
+    StepResponse stepResponse =
+        serviceStepV3Helper.handleFreezeResponse(ambiance, environmentOutcome, OutcomeExpressionConstants.ENVIRONMENT);
+    if (stepResponse != null) {
+      return stepResponse;
+    }
+
+    stepResponse = SdkCoreStepUtils.createStepResponseFromChildResponse(responseDataMap);
 
     final NGLogCallback logCallback = getLogCallback(ambiance, false, ENVIRONMENT_STEP_COMMAND_UNIT);
 
@@ -222,9 +250,6 @@ public class CustomStageEnvironmentStep implements ChildrenExecutable<CustomStag
                                         .setEndTime(System.currentTimeMillis())
                                         .build();
     }
-
-    final EnvironmentOutcome environmentOutcome = (EnvironmentOutcome) sweepingOutputService.resolve(
-        ambiance, RefObjectUtils.getOutcomeRefObject(OutputExpressionConstants.ENVIRONMENT));
 
     stepOutcomes.add(StepResponse.StepOutcome.builder()
                          .name(OutcomeExpressionConstants.ENVIRONMENT)
