+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/service/steps/ServiceStepV3.java
@@ -18,7 +18,6 @@ import static io.harness.cdng.service.steps.constants.ServiceStepConstants.SERVI
 import static io.harness.data.structure.CollectionUtils.emptyIfNull;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
-import static io.harness.eraro.ErrorCode.FREEZE_EXCEPTION;
 
 import static java.lang.String.format;
 
@@ -35,10 +34,8 @@ import io.harness.cdng.environment.helper.EnvironmentInfraFilterHelper;
 import io.harness.cdng.environment.helper.EnvironmentMapper;
 import io.harness.cdng.environment.helper.EnvironmentPlanCreatorHelper;
 import io.harness.cdng.environment.helper.EnvironmentStepsUtils;
-import io.harness.cdng.freeze.FreezeOutcome;
 import io.harness.cdng.gitops.steps.EnvClusterRefs;
 import io.harness.cdng.gitops.steps.GitOpsEnvOutCome;
-import io.harness.cdng.helpers.NgExpressionHelper;
 import io.harness.cdng.hooks.steps.ServiceHooksOutcome;
 import io.harness.cdng.k8s.HarnessRelease;
 import io.harness.cdng.service.ServiceSpec;
@@ -53,16 +50,10 @@ import io.harness.cdng.stepsdependency.constants.OutcomeExpressionConstants;
 import io.harness.cdng.visitor.YamlTypes;
 import io.harness.data.structure.EmptyPredicate;
 import io.harness.encryption.Scope;
-import io.harness.eraro.Level;
 import io.harness.exception.InvalidRequestException;
 import io.harness.exception.UnresolvedExpressionsException;
 import io.harness.exception.WingsException;
 import io.harness.freeze.beans.FreezeEntityType;
-import io.harness.freeze.beans.response.FreezeSummaryResponseDTO;
-import io.harness.freeze.helpers.FreezeRBACHelper;
-import io.harness.freeze.notifications.NotificationHelper;
-import io.harness.freeze.service.FreezeEvaluateService;
-import io.harness.freeze.service.FrozenExecutionService;
 import io.harness.gitx.GitXTransientBranchGuard;
 import io.harness.logging.CommandExecutionStatus;
 import io.harness.logging.LogLevel;
@@ -87,15 +78,9 @@ import io.harness.ng.core.serviceoverridev2.beans.ServiceOverridesType;
 import io.harness.ng.core.utils.ServiceOverrideV2ValidationHelper;
 import io.harness.pms.contracts.ambiance.Ambiance;
 import io.harness.pms.contracts.execution.ChildrenExecutableResponse;
-import io.harness.pms.contracts.execution.Status;
-import io.harness.pms.contracts.execution.failure.FailureData;
-import io.harness.pms.contracts.execution.failure.FailureInfo;
-import io.harness.pms.contracts.execution.failure.FailureType;
-import io.harness.pms.contracts.plan.ExpressionMode;
 import io.harness.pms.contracts.steps.StepCategory;
 import io.harness.pms.execution.utils.AmbianceUtils;
 import io.harness.pms.execution.utils.StatusUtils;
-import io.harness.pms.expression.EngineExpressionService;
 import io.harness.pms.merger.helpers.MergeHelper;
 import io.harness.pms.sdk.core.data.OptionalSweepingOutput;
 import io.harness.pms.sdk.core.resolver.RefObjectUtils;
@@ -106,14 +91,12 @@ import io.harness.pms.sdk.core.steps.io.StepResponse;
 import io.harness.pms.security.PmsSecurityContextEventGuard;
 import io.harness.pms.yaml.ParameterField;
 import io.harness.pms.yaml.YamlUtils;
-import io.harness.rbac.CDNGRbacUtility;
 import io.harness.steps.OutputExpressionConstants;
 import io.harness.steps.SdkCoreStepUtils;
 import io.harness.steps.StepUtils;
 import io.harness.steps.environment.EnvironmentOutcome;
 import io.harness.tasks.ResponseData;
 import io.harness.utils.IdentifierRefHelper;
-import io.harness.utils.NGFeatureFlagHelperService;
 import io.harness.utils.YamlPipelineUtils;
 import io.harness.yaml.core.variables.NGVariable;
 import io.harness.yaml.core.variables.SecretNGVariable;
@@ -122,14 +105,12 @@ import io.harness.yaml.utils.NGVariablesUtils;
 import software.wings.beans.LogColor;
 import software.wings.beans.LogHelper;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.List;
@@ -153,12 +134,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
   @Inject private EnvironmentService environmentService;
   @Inject private ServiceOverrideService serviceOverrideService;
   @Inject private ServiceStepOverrideHelper serviceStepOverrideHelper;
-  @Inject private FreezeEvaluateService freezeEvaluateService;
-  @Inject private FrozenExecutionService frozenExecutionService;
-  @Inject private NGFeatureFlagHelperService ngFeatureFlagHelperService;
-  @Inject private NotificationHelper notificationHelper;
-  @Inject private EngineExpressionService engineExpressionService;
-  @Inject NgExpressionHelper ngExpressionHelper;
   @Inject private EnvironmentGroupService environmentGroupService;
   @Inject private EnvironmentInfraFilterHelper environmentInfraFilterHelper;
   @Inject @Named("PRIVILEGED") private AccessControlClient accessControlClient;
@@ -212,7 +187,8 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
             ambiance, stepParameters, servicePartResponse, serviceStepLogCallback, entityMap, overrideLogCallback);
       }
 
-      ChildrenExecutableResponse childrenExecutableResponse = executeFreezePart(ambiance, entityMap);
+      ChildrenExecutableResponse childrenExecutableResponse = serviceStepV3Helper.executeFreezePart(
+          ambiance, entityMap, List.of(SERVICE_STEP_COMMAND_UNIT, OVERRIDES_COMMAND_UNIT));
       if (childrenExecutableResponse != null) {
         return childrenExecutableResponse;
       }
@@ -619,7 +595,13 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
     }
     final NGServiceV2InfoConfig ngServiceV2InfoConfig = ngServiceConfig.getNgServiceV2InfoConfig();
 
-    StepResponse stepResponse = handleFreezeResponse(ambiance, ngServiceV2InfoConfig, stepParameters);
+    ServiceStepOutcome serviceStepOutcome = ServiceStepOutcome.fromServiceStepV2(
+        stepParameters.getServiceRef().getValue().trim(), ngServiceV2InfoConfig.getName(),
+        ngServiceV2InfoConfig.getServiceDefinition().getType().getYamlName(), ngServiceV2InfoConfig.getDescription(),
+        ngServiceV2InfoConfig.getTags(), ngServiceV2InfoConfig.getGitOpsEnabled());
+
+    StepResponse stepResponse =
+        serviceStepV3Helper.handleFreezeResponse(ambiance, serviceStepOutcome, OutcomeExpressionConstants.SERVICE);
     if (stepResponse != null) {
       return stepResponse;
     }
@@ -804,90 +786,5 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
     serviceInputsYaml.put(YamlTypes.SERVICE_ENTITY, serviceInputs);
     return MergeHelper.mergeRuntimeInputValuesAndCheckForRuntimeInOriginalYaml(
         originalServiceYaml, YamlPipelineUtils.writeYamlString(serviceInputsYaml), true, true);
-  }
-
-  @VisibleForTesting
-  protected ChildrenExecutableResponse executeFreezePart(
-      Ambiance ambiance, Map<FreezeEntityType, List<String>> entityMap) {
-    String accountId = AmbianceUtils.getAccountId(ambiance);
-    String orgId = AmbianceUtils.getOrgIdentifier(ambiance);
-    String projectId = AmbianceUtils.getProjectIdentifier(ambiance);
-    if (FreezeRBACHelper.checkIfUserHasFreezeOverrideAccess(ngFeatureFlagHelperService, accountId, orgId, projectId,
-            accessControlClient, CDNGRbacUtility.constructPrincipalFromAmbiance(ambiance))) {
-      return null;
-    }
-    List<FreezeSummaryResponseDTO> globalFreezeConfigs;
-    List<FreezeSummaryResponseDTO> manualFreezeConfigs;
-    globalFreezeConfigs = freezeEvaluateService.anyGlobalFreezeActive(accountId, orgId, projectId);
-    manualFreezeConfigs = freezeEvaluateService.getActiveManualFreezeEntities(accountId, orgId, projectId, entityMap);
-    if (globalFreezeConfigs.size() + manualFreezeConfigs.size() > 0) {
-      log.info("Deployment Freeze is Active for the given service.");
-      sweepingOutputService.consume(ambiance, ServiceStepV3Constants.FREEZE_SWEEPING_OUTPUT,
-          FreezeOutcome.builder()
-              .frozen(true)
-              .manualFreezeConfigs(manualFreezeConfigs)
-              .globalFreezeConfigs(globalFreezeConfigs)
-              .build(),
-          "");
-      log.info("Adding Children as empty.");
-      String executionUrl =
-          engineExpressionService.renderExpression(ambiance, ServiceStepV3Constants.PIPELINE_EXECUTION_EXPRESSION,
-              ExpressionMode.RETURN_ORIGINAL_EXPRESSION_IF_UNRESOLVED);
-      String baseUrl = ngExpressionHelper.getBaseUrl(AmbianceUtils.getAccountId(ambiance));
-      notificationHelper.sendNotificationForFreezeConfigs(
-          manualFreezeConfigs, globalFreezeConfigs, ambiance, executionUrl, baseUrl);
-      return ChildrenExecutableResponse.newBuilder()
-          .addAllLogKeys(emptyIfNull(StepUtils.generateLogKeys(
-              StepUtils.generateLogAbstractions(ambiance), List.of(SERVICE_STEP_COMMAND_UNIT, OVERRIDES_COMMAND_UNIT))))
-          .addAllChildren(Collections.emptyList())
-          .addAllUnits(List.of(SERVICE_STEP_COMMAND_UNIT, OVERRIDES_COMMAND_UNIT))
-          .build();
-    }
-    return null;
-  }
-
-  @VisibleForTesting
-  protected StepResponse handleFreezeResponse(
-      Ambiance ambiance, NGServiceV2InfoConfig ngServiceV2InfoConfig, ServiceStepV3Parameters serviceStepParameters) {
-    final List<StepResponse.StepOutcome> stepOutcomes = new ArrayList<>();
-    final OptionalSweepingOutput freezeOutcomeOptional = sweepingOutputService.resolveOptional(
-        ambiance, RefObjectUtils.getOutcomeRefObject(ServiceStepV3Constants.FREEZE_SWEEPING_OUTPUT));
-
-    if (freezeOutcomeOptional.isFound()) {
-      FreezeOutcome freezeOutcome = (FreezeOutcome) freezeOutcomeOptional.getOutput();
-      if (freezeOutcome.isFrozen()) {
-        frozenExecutionService.createFrozenExecution(
-            ambiance, freezeOutcome.getManualFreezeConfigs(), freezeOutcome.getGlobalFreezeConfigs());
-
-        stepOutcomes.add(StepResponse.StepOutcome.builder()
-                             .name(OutcomeExpressionConstants.FREEZE_OUTCOME)
-                             .outcome(freezeOutcome)
-                             .group(StepCategory.STAGE.name())
-                             .build());
-        stepOutcomes.add(
-            StepResponse.StepOutcome.builder()
-                .name(OutcomeExpressionConstants.SERVICE)
-                .outcome(ServiceStepOutcome.fromServiceStepV2(serviceStepParameters.getServiceRef().getValue().trim(),
-                    ngServiceV2InfoConfig.getName(),
-                    ngServiceV2InfoConfig.getServiceDefinition().getType().getYamlName(),
-                    ngServiceV2InfoConfig.getDescription(), ngServiceV2InfoConfig.getTags(),
-                    ngServiceV2InfoConfig.getGitOpsEnabled()))
-                .group(StepCategory.STAGE.name())
-                .build());
-        return StepResponse.builder()
-            .stepOutcomes(stepOutcomes)
-            .failureInfo(FailureInfo.newBuilder()
-                             .addFailureData(FailureData.newBuilder()
-                                                 .addFailureTypes(FailureType.FREEZE_ACTIVE_FAILURE)
-                                                 .setLevel(Level.ERROR.name())
-                                                 .setCode(FREEZE_EXCEPTION.name())
-                                                 .setMessage("Pipeline Aborted due to freeze")
-                                                 .build())
-                             .build())
-            .status(Status.FREEZE_FAILED)
-            .build();
-      }
-    }
-    return null;
   }
 }
