+++ 125-cd-nextgen/src/test/java/io/harness/cdng/service/steps/ServiceStepV3Test.java
@@ -19,13 +19,10 @@ import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import io.harness.CategoryTest;
 import io.harness.accesscontrol.acl.api.AccessCheckResponseDTO;
 import io.harness.accesscontrol.acl.api.Principal;
-import io.harness.accesscontrol.acl.api.Resource;
-import io.harness.accesscontrol.acl.api.ResourceScope;
 import io.harness.accesscontrol.clients.AccessControlClient;
 import io.harness.beans.common.VariablesSweepingOutput;
 import io.harness.category.element.UnitTests;
@@ -33,7 +30,6 @@ import io.harness.cdng.artifact.bean.yaml.DockerHubArtifactConfig;
 import io.harness.cdng.artifact.outcome.ArtifactsOutcome;
 import io.harness.cdng.configfile.ConfigFilesOutcome;
 import io.harness.cdng.expressions.CDExpressionResolver;
-import io.harness.cdng.freeze.FreezeOutcome;
 import io.harness.cdng.gitops.steps.GitOpsEnvOutCome;
 import io.harness.cdng.helpers.NgExpressionHelper;
 import io.harness.cdng.manifest.steps.outcome.ManifestsOutcome;
@@ -51,15 +47,7 @@ import io.harness.encryption.SecretRefData;
 import io.harness.exception.InvalidRequestException;
 import io.harness.exception.UnresolvedExpressionsException;
 import io.harness.freeze.beans.FreezeEntityType;
-import io.harness.freeze.beans.FreezeStatus;
-import io.harness.freeze.beans.FreezeType;
-import io.harness.freeze.beans.response.FreezeSummaryResponseDTO;
-import io.harness.freeze.beans.yaml.FreezeConfig;
-import io.harness.freeze.beans.yaml.FreezeInfoConfig;
-import io.harness.freeze.entity.FreezeConfigEntity;
-import io.harness.freeze.mappers.NGFreezeDtoMapper;
 import io.harness.freeze.notifications.NotificationHelper;
-import io.harness.freeze.service.FreezeEvaluateService;
 import io.harness.ng.core.dto.ResponseDTO;
 import io.harness.ng.core.environment.beans.Environment;
 import io.harness.ng.core.environment.beans.EnvironmentType;
@@ -108,10 +96,8 @@ import com.google.common.collect.Lists;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.EnumMap;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -138,7 +124,6 @@ public class ServiceStepV3Test extends CategoryTest {
   @Mock private CDExpressionResolver expressionResolver;
   @Mock private ServiceStepOverrideHelper serviceStepOverrideHelper;
   @Mock private NGFeatureFlagHelperService ngFeatureFlagHelperService;
-  @Mock private FreezeEvaluateService freezeEvaluateService;
   @Mock private AccessControlClient accessControlClient;
   @Mock private NotificationHelper notificationHelper;
   @Mock private EngineExpressionService engineExpressionService;
@@ -179,6 +164,8 @@ public class ServiceStepV3Test extends CategoryTest {
     doNothing()
         .when(serviceCustomSweepingOutputHelper)
         .saveAdditionalServiceFieldsToSweepingOutput(any(NGServiceConfig.class), any(Ambiance.class));
+
+    doReturn(null).when(serviceStepV3Helper).executeFreezePart(any(), any(), any());
   }
   @After
   public void tearDown() throws Exception {
@@ -356,7 +343,6 @@ public class ServiceStepV3Test extends CategoryTest {
     final ServiceEntity serviceEntity = testServiceEntity();
     final Environment environment = testEnvEntity();
 
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
     doReturn(false).when(overrideV2ValidationHelper).isOverridesV2Enabled(any(), any(), any());
     mockService(serviceEntity);
     mockEnv(environment);
@@ -372,7 +358,7 @@ public class ServiceStepV3Test extends CategoryTest {
     assertThat(response.getLogKeysCount()).isEqualTo(2);
 
     ArgumentCaptor<Map> captor = ArgumentCaptor.forClass(Map.class);
-    verify(freezeEvaluateService, times(1)).getActiveManualFreezeEntities(any(), any(), any(), captor.capture());
+    verify(serviceStepV3Helper, times(1)).executeFreezePart(any(), captor.capture(), any());
 
     Map<FreezeEntityType, List<String>> entityMap = captor.getValue();
     assertThat(entityMap.size()).isEqualTo(6);
@@ -389,7 +375,6 @@ public class ServiceStepV3Test extends CategoryTest {
     final ServiceEntity serviceEntity = testOrgServiceEntity();
     final Environment environment = testOrgEnvEntity();
 
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
     doReturn(false).when(overrideV2ValidationHelper).isOverridesV2Enabled(any(), any(), any());
 
     mockService(serviceEntity);
@@ -406,7 +391,7 @@ public class ServiceStepV3Test extends CategoryTest {
     assertThat(response.getLogKeysCount()).isEqualTo(2);
 
     ArgumentCaptor<Map> captor = ArgumentCaptor.forClass(Map.class);
-    verify(freezeEvaluateService, times(1)).getActiveManualFreezeEntities(any(), any(), any(), captor.capture());
+    verify(serviceStepV3Helper, times(1)).executeFreezePart(any(), captor.capture(), any());
 
     Map<FreezeEntityType, List<String>> entityMap = captor.getValue();
     assertThat(entityMap.size()).isEqualTo(6);
@@ -423,7 +408,6 @@ public class ServiceStepV3Test extends CategoryTest {
     final ServiceEntity serviceEntity = testAccountServiceEntity();
     final Environment environment = testAccountEnvEntity();
 
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
     doReturn(false).when(overrideV2ValidationHelper).isOverridesV2Enabled(any(), any(), any());
     mockService(serviceEntity);
     mockEnv(environment);
@@ -439,7 +423,7 @@ public class ServiceStepV3Test extends CategoryTest {
     assertThat(response.getLogKeysCount()).isEqualTo(2);
 
     ArgumentCaptor<Map> captor = ArgumentCaptor.forClass(Map.class);
-    verify(freezeEvaluateService, times(1)).getActiveManualFreezeEntities(any(), any(), any(), captor.capture());
+    verify(serviceStepV3Helper, times(1)).executeFreezePart(any(), captor.capture(), any());
 
     Map<FreezeEntityType, List<String>> entityMap = captor.getValue();
     assertThat(entityMap.size()).isEqualTo(6);
@@ -1049,96 +1033,6 @@ public class ServiceStepV3Test extends CategoryTest {
     assertThat(stepResponse.getStatus()).isEqualTo(Status.FAILED);
   }
 
-  @Test
-  @Owner(developers = OwnerRule.ABHINAV_MITTAL)
-  @Category(UnitTests.class)
-  public void testExecuteFreezePart() {
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
-    List<FreezeSummaryResponseDTO> freezeSummaryResponseDTOList = Lists.newArrayList(createGlobalFreezeResponse());
-    doReturn(freezeSummaryResponseDTOList)
-        .when(freezeEvaluateService)
-        .anyGlobalFreezeActive(anyString(), anyString(), anyString());
-    when(
-        accessControlClient.hasAccess(any(Principal.class), any(ResourceScope.class), any(Resource.class), anyString()))
-        .thenReturn(false);
-    Map<FreezeEntityType, List<String>> entityMap = new HashMap<>();
-
-    ChildrenExecutableResponse childrenExecutableResponse = step.executeFreezePart(buildAmbiance(), entityMap);
-    ArgumentCaptor<ExecutionSweepingOutput> captor = ArgumentCaptor.forClass(ExecutionSweepingOutput.class);
-    ArgumentCaptor<String> stringCaptor = ArgumentCaptor.forClass(String.class);
-
-    verify(sweepingOutputService, times(1))
-        .consume(any(Ambiance.class), stringCaptor.capture(), captor.capture(), anyString());
-
-    List<ExecutionSweepingOutput> allValues = captor.getAllValues();
-
-    Map<Class, ExecutionSweepingOutput> outputsMap =
-        allValues.stream().collect(Collectors.toMap(ExecutionSweepingOutput::getClass, a -> a));
-    FreezeOutcome freezeOutcome = (FreezeOutcome) outputsMap.get(FreezeOutcome.class);
-    assertThat(freezeOutcome.isFrozen()).isEqualTo(true);
-    assertThat(freezeOutcome.getGlobalFreezeConfigs()).isEqualTo(freezeSummaryResponseDTOList);
-    assertThat(childrenExecutableResponse.getChildrenCount()).isEqualTo(0);
-  }
-
-  @Test
-  @Owner(developers = OwnerRule.ABHINAV_MITTAL)
-  @Category(UnitTests.class)
-  public void testExecuteFreezePartWithEmptyFreezeList() {
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
-    List<FreezeSummaryResponseDTO> freezeSummaryResponseDTOList = Collections.emptyList();
-    doReturn(freezeSummaryResponseDTOList)
-        .when(freezeEvaluateService)
-        .anyGlobalFreezeActive(anyString(), anyString(), anyString());
-    when(
-        accessControlClient.hasAccess(any(Principal.class), any(ResourceScope.class), any(Resource.class), anyString()))
-        .thenReturn(false);
-    Map<FreezeEntityType, List<String>> entityMap = new HashMap<>();
-
-    ChildrenExecutableResponse childrenExecutableResponse = step.executeFreezePart(buildAmbiance(), entityMap);
-    ArgumentCaptor<ExecutionSweepingOutput> captor = ArgumentCaptor.forClass(ExecutionSweepingOutput.class);
-    ArgumentCaptor<String> stringCaptor = ArgumentCaptor.forClass(String.class);
-
-    verify(sweepingOutputService, times(0))
-        .consume(any(Ambiance.class), stringCaptor.capture(), captor.capture(), anyString());
-    assertThat(childrenExecutableResponse).isNull();
-  }
-
-  @Test
-  @Owner(developers = OwnerRule.ABHINAV_MITTAL)
-  @Category(UnitTests.class)
-  public void testExecuteFreezePartIfOverrideFreeze() {
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
-    List<FreezeSummaryResponseDTO> freezeSummaryResponseDTOList = Lists.newArrayList(createGlobalFreezeResponse());
-    doReturn(freezeSummaryResponseDTOList)
-        .when(freezeEvaluateService)
-        .anyGlobalFreezeActive(anyString(), anyString(), anyString());
-    when(
-        accessControlClient.hasAccess(any(Principal.class), any(ResourceScope.class), any(Resource.class), anyString()))
-        .thenReturn(true);
-    Map<FreezeEntityType, List<String>> entityMap = new HashMap<>();
-
-    ChildrenExecutableResponse childrenExecutableResponse = step.executeFreezePart(buildAmbiance(), entityMap);
-    assertThat(childrenExecutableResponse).isEqualTo(null);
-  }
-
-  @Test
-  @Owner(developers = OwnerRule.YOGESH)
-  @Category(UnitTests.class)
-  public void testExecuteFreezePartIfFreezeNotActive() {
-    doReturn(true).when(ngFeatureFlagHelperService).isEnabled(anyString(), any());
-    doReturn(new LinkedList<>())
-        .when(freezeEvaluateService)
-        .anyGlobalFreezeActive(anyString(), anyString(), anyString());
-    Map<FreezeEntityType, List<String>> entityMap = new HashMap<>();
-
-    step.executeFreezePart(buildAmbiance(), entityMap);
-    ArgumentCaptor<ExecutionSweepingOutput> captor = ArgumentCaptor.forClass(ExecutionSweepingOutput.class);
-    ArgumentCaptor<String> stringCaptor = ArgumentCaptor.forClass(String.class);
-
-    verify(sweepingOutputService, times(0))
-        .consume(any(Ambiance.class), stringCaptor.capture(), captor.capture(), anyString());
-  }
-
   private ServiceEntity testServiceEntity() {
     final String serviceYaml = "service:\n"
         + "  name: service-name\n"
@@ -1438,20 +1332,6 @@ public class ServiceStepV3Test extends CategoryTest {
     doReturn(Optional.ofNullable(serviceOverrides))
         .when(serviceOverrideService)
         .get(anyString(), anyString(), anyString(), anyString(), anyString());
-  }
-
-  private FreezeSummaryResponseDTO createGlobalFreezeResponse() {
-    FreezeConfig freezeConfig = FreezeConfig.builder()
-                                    .freezeInfoConfig(FreezeInfoConfig.builder()
-                                                          .identifier("_GLOBAL_")
-                                                          .name("Global Freeze")
-                                                          .status(FreezeStatus.DISABLED)
-                                                          .build())
-                                    .build();
-    String yaml = NGFreezeDtoMapper.toYaml(freezeConfig);
-    FreezeConfigEntity freezeConfigEntity = NGFreezeDtoMapper.toFreezeConfigEntity(
-        "accountId", "orgIdentifier", "projectIdentifier", yaml, FreezeType.GLOBAL);
-    return NGFreezeDtoMapper.prepareFreezeResponseSummaryDto(freezeConfigEntity);
   }
 
   private static EnumMap<ServiceOverridesType, NGServiceOverrideConfigV2> getOverridesConfigsV2MapTestData(
